Changed initial handshake protocol
Reduced number of phases of initial protocol and some overhead on
messages.
Now the !setnick message is sent by the ServerConnection *before*
starting the receiveMsg loop, thus avoiding race conditions between
it and the !useradd messages to add the users list.
Also, !useradd is now capable of adding several users with the same
message, thus the Server only sends 1 !useradd when the handshake ends.
The role parameter in !useradd is no longer optional, and the message
must contain an even number of parameters (user role, user role, ...)